phantom usage is a dumb idea..
Functions are automatically curried. Hence, there is no way to
support phantom usage unless we track lifetime as well. The motivation
of phantom usage is to simplify using linear resources: to free users
from passing in a resource and collecting the returned resource again.
Without phantom usage, the alternative is to support some syntactical
sugar for such common use cases, along the lines of "do notations".